requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
django-import-export==2.1.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
psycopg2==2.8.5
pytz==2019.3
PyYAML==5.3.1
sqlparse==0.3.1
tablib==1.1.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
